Move APISDGFIP and APISFRANCECONNECTED into dataProviderConfigurations File
- When we need to add a new FC Apis or DGFiP Apis, we could update the login page and welcome message from only one file.
